requirements.txt
Django==3.0.6
virtualenv==20.0.4
pyrebase4==4.4.3
gunicorn-20.1.0

